Development

+++ BackEnd/CmApp/Services/CarService.cs
@@ -5,6 +5,7 @@ using CmApp.Utils;
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace CmApp.Services
@@ -69,6 +70,8 @@ namespace CmApp.Services
                 carEntity.BodyType = "Coupe";
             else if (carEntity.BodyType == "SAV")
                 carEntity.BodyType = "SUV";
+            else if (carEntity.BodyType == "LIM")
+                carEntity.BodyType = "Limousine";
 
 
             var eqResults = WebScraper.GetVehicleEquipment(car.Vin, car.Make);
@@ -79,6 +82,7 @@ namespace CmApp.Services
 
             carEntity.Equipment = equipment;
             carEntity.Vin = car.Vin.ToUpper();
+            carEntity.MainImageUrl = Settings.DefaultImageUrl;
 
             //inserting vehicle data
             var insertedCar = await CarRepository.InsertCar(carEntity);
@@ -96,10 +100,17 @@ namespace CmApp.Services
                     var imgName = insertedCar.Id + "_image" + count + "." + imageType;
 
                     var bytes = FileRepository.Base64ToByteArray(base64);
-                    await CarRepository.UploadImageToCar(insertedCar.Id, bytes, imgName);
+                    var res = await CarRepository.UploadImageToCar(insertedCar.Id, bytes, imgName);
+                    if (count == 1)
+                    {
+                        var url = await FileRepository.GetFileUrl(res.Key);
+                        await CarRepository.UpdateCarMainImg(insertedCar.Id, url);
+                    }
+                    
                     count++;
                 }
             }
+
            /* else
             {
                 var defaultImg = await FileRepository.GetFile(Settings.DefaultImage);
@@ -165,6 +176,7 @@ namespace CmApp.Services
             carEntity.Equipment = equipment;
             carEntity.Vin = car.Vin.ToUpper();
 
+            carEntity.MainImageUrl = Settings.DefaultImageUrl;
             //inserting vehicle data
             var insertedCar = await CarRepository.InsertCar(carEntity);
 
@@ -181,7 +193,12 @@ namespace CmApp.Services
                     var imgName = insertedCar.Id + "_image" + count + "." + imageType;
 
                     var bytes = FileRepository.Base64ToByteArray(base64);
-                    await CarRepository.UploadImageToCar(insertedCar.Id, bytes, imgName);
+                    var res = await CarRepository.UploadImageToCar(insertedCar.Id, bytes, imgName);
+                    if (count == 1)
+                    {
+                        var url = await FileRepository.GetFileUrl(res.Key);
+                        await CarRepository.UpdateCarMainImg(insertedCar.Id, url);
+                    }
                     count++;
                 }
             }
@@ -205,6 +222,12 @@ namespace CmApp.Services
                 throw new BusinessException("Can not insert car, because car is empty!");
 
             car.Vin = car.Vin.ToUpper();
+            string mainImg = "";
+            if (car.Base64images.Count > 0)
+                mainImg = car.Base64images[0];
+            else
+                mainImg = Settings.DefaultImageUrl;
+            car.MainImageUrl = mainImg;
             //inserting vehicle data
             var insertedCar = await CarRepository.InsertCar(car);
 
@@ -221,7 +244,12 @@ namespace CmApp.Services
                     var imgName = insertedCar.Id + "_image" + count + "." + imageType;
 
                     var bytes = FileRepository.Base64ToByteArray(base64);
-                    await CarRepository.UploadImageToCar(insertedCar.Id, bytes, imgName);
+                    var res = await CarRepository.UploadImageToCar(insertedCar.Id, bytes, imgName);
+                    if (count == 1)
+                    {
+                        var url = await FileRepository.GetFileUrl(res.Key);
+                        await CarRepository.UpdateCarMainImg(insertedCar.Id, url);
+                    }
                     count++;
                 }
             }
@@ -241,6 +269,11 @@ namespace CmApp.Services
 
         public async Task<CarEntity> InsertCar(CarEntity car)
         {
+            var userCars = await GetAllUserCars(car.User);
+            var userVins = userCars.Select(x => x.Vin).ToList();
+            if (userVins.Contains(car.Vin))
+                throw new BusinessException("There is already a car with this VIN number");
+
             if (car.Make == null || car.Make == "")
                 throw new BusinessException("Make not defined");
             if (car.Make == "BMW" && car.Model == "")
@@ -274,7 +307,7 @@ namespace CmApp.Services
 
                     var url = await FileRepository.GetFileUrl(fileId);
 
-                    car.MainImgUrl = url;
+                    car.MainImageUrl = url;
                 }
                 else
                 {
@@ -286,10 +319,10 @@ namespace CmApp.Services
 
                         var url = await FileRepository.GetFileUrl(fileId);
 
-                        car.MainImgUrl = url;
+                        car.MainImageUrl = url;
                     }
                     else
-                        car.MainImgUrl = Settings.DefaultImageUrl;
+                        car.MainImageUrl = Settings.DefaultImageUrl;
                 }
                     
             }
@@ -301,6 +334,9 @@ namespace CmApp.Services
         {
             if (car.User != userId)
                 throw new BusinessException("Car does not exist");
+            if(car.Equipment.Count != car.Equipment.Distinct().Count())
+                throw new BusinessException("Car cannot have multiple equipment with the same code!");
+
             await CarRepository.UpdateCar(carId, car);
             await CarRepository.DeleteAllCarImages(carId);
 
@@ -329,7 +365,7 @@ namespace CmApp.Services
             var car = await CarRepository.GetCarById(id);
             if (car == null)
                 throw new BusinessException("Car with provided id does not exists!");
-            car.MainImgUrl = Settings.DefaultImageUrl;
+            car.MainImageUrl = Settings.DefaultImageUrl;
 
             return car;
         }
